Adding mechanism for plugable JS engines
